fix(frontend): escape user content in JSON-LD via shared serializer (xss)
