+++ frontend/app/(marketing)/_components/breadcrumbJsonLd.ts
@@ -1,23 +1,10 @@
-// Server-side breadcrumb JSON-LD helper. Mirrors
-// src/utils/structuredData.ts#generateBreadcrumbStructuredData but reads the
-// origin from NEXT_PUBLIC_SITE_ORIGIN since we have no window context.
+// Server-side breadcrumb JSON-LD helper. Delegates serialization to the
+// shared lib/seo/jsonLd helper so the inline-script escape contract has
+// exactly one home.
+import { jsonLdHtml } from '../../../lib/seo/jsonLd'
 
 const SITE_ORIGIN =
   process.env.NEXT_PUBLIC_SITE_ORIGIN || 'https://webwhen.ai'
-
-const LS = String.fromCharCode(0x2028)
-const PS = String.fromCharCode(0x2029)
-
-function escapeForScriptTag(json: string): string {
-  return json
-    .replace(/\\/g, '\\\\')
-    .replace(/</g, '\\u003c')
-    .replace(/\//g, '\\/')
-    .split(LS)
-    .join('\\u2028')
-    .split(PS)
-    .join('\\u2029')
-}
 
 export interface BreadcrumbItem {
   name: string
@@ -35,5 +22,5 @@ export function breadcrumbJsonLd(items: BreadcrumbItem[]): string {
       item: `${SITE_ORIGIN}${item.path}`,
     })),
   }
-  return escapeForScriptTag(JSON.stringify(data))
+  return jsonLdHtml(data)
 }
